backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
joblib==1.4.2
scikit-learn==1.7.2
numpy==2.2.1
pandas==2.2.3
aiohttp==3.11.11
beautifulsoup4==4.12.3
nltk==3.9.1
python-multipart==0.0.20
lxml==5.3.0

# Advanced AI features
transformers==4.47.1
sentence-transformers==3.3.1
torch==2.5.1
faiss-cpu==1.9.0
openai==1.59.6
